requirements.txt
Django==2.0.6
django-appconf==1.0.2
django-bootstrap3==11.0.0
django-cleanup==2.1.0
django-js-asset==1.1.0
django-mptt==0.9.1
django-summernote==0.8.10.0
djangorestframework==3.9.0
easy-thumbnails==2.5
mysqlclient==1.3.13
pilkit==2.0
Pillow==5.1.0
pytz==2018.5
requests==2.6.0
six==1.9.0
